+++ sqlparser/CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 0.15.1-dev
 
 - New analysis checks for `RETURNING`: Disallow `table.*` syntax and aggregate expressions
+- Fix resolving columns when `RETURNING` is used in an `UPDATE FROM` statement
 
 ## 0.15.0